requirements.txt
# MCP Image Validator Requirements
# Python 3.10+

# Core dependencies
mcp>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Optional for development
# websockets>=12.0  # If using WebSocket transport
